Updated to refactored versions
