+++ code/backend/src/main/java/com/cs673/careerforge/exceptions/RestExceptionHandler.java
@@ -1,10 +1,9 @@
-package com.cs673.careerforge.exception;
+package com.cs673.careerforge.exceptions;
 
 import jakarta.validation.ConstraintViolationException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.util.StringUtils;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
